LANGUAGEとUSER_IDをgにセット
