+++ .devcontainer/docker-compose.yml
@@ -80,9 +80,10 @@ services:
         - HTTPS_PROXY=${HTTPS_PROXY}
         - https_proxy=${HTTPS_PROXY}
     environment:
-      - EXECUTE_INTERVAL=10
+      - EXECUTE_INTERVAL=1
       - PYTHONPATH=/exastro/
-      - DEFAULT_LANGUAGE=en
+      - LANGUAGE=en
+      - USER_ID=20101
       - DB_HOST=ita-mariadb
       - DB_PORT=3306
       - DB_DATADBASE=ITA_DB
@@ -92,53 +93,51 @@ services:
       - STORAGEPATH=/storage/
       - HOST_STORAGEPATH=/home/almalinux/uekawa-tomonari/it-automation2-test/.volumes/storage/
       - CONTAINER_BASE=docker
-      - LAST_UPDATE_USER_ID=3
-  # ita-by-ansible-legacy-role-vars-listup:
-  #   build:
-  #     dockerfile: ./ita_root/ita_by_ansible_legacy_role_vars_listup/Dockerfile
-  #     context: ../.
-  #     args:
-  #       - HTTP_PROXY=${HTTP_PROXY}
-  #       - http_proxy=${HTTP_PROXY}
-  #       - HTTPS_PROXY=${HTTPS_PROXY}
-  #       - https_proxy=${HTTPS_PROXY}
-  #   environment:
-  #     - EXECUTE_INTERVAL=10
-  #     - PYTHONPATH=/exastro/
-  #     - DEFAULT_LANGUAGE=en
-  #     - DB_HOST=ita-mariadb
-  #     - DB_PORT=3306
-  #     - DB_DATADBASE=ITA_DB
-  #     - DB_USER=ITA_USER
-  #     - DB_PASSWORD=password
-  #     - DB_ROOT_PASSWORD=password
-  #     - STORAGEPATH=/storage/
-  #     - GITLAB_HOST=10.197.18.183
-  #     - GITLAB_PORT=40080
-  #     - GITLAB_ROOT_TOKEN=mzsQj8VU_zcSoSZsSX__
-  # ita-by-ansible-towermaster-sync:
-  #   build:
-  #     dockerfile: ./ita_root/ita_by_ansible_towermaster_sync/Dockerfile
-  #     context: ../.
-  #     args:
-  #       - HTTP_PROXY=${HTTP_PROXY}
-  #       - http_proxy=${HTTP_PROXY}
-  #       - HTTPS_PROXY=${HTTPS_PROXY}
-  #       - https_proxy=${HTTPS_PROXY}
-  #   environment:
-  #     - EXECUTE_INTERVAL=10
-  #     - PYTHONPATH=/exastro/
-  #     - DEFAULT_LANGUAGE=en
-  #     - DB_HOST=ita-mariadb
-  #     - DB_PORT=3306
-  #     - DB_DATADBASE=ITA_DB
-  #     - DB_USER=ITA_USER
-  #     - DB_PASSWORD=password
-  #     - DB_ROOT_PASSWORD=password
-  #     - STORAGEPATH=/storage/
-  #     - GITLAB_HOST=10.197.18.183
-  #     - GITLAB_PORT=40080
-  #     - GITLAB_ROOT_TOKEN=mzsQj8VU_zcSoSZsSX__
+      - GITLAB_HOST=10.197.18.183
+      - GITLAB_PORT=40080
+      - GITLAB_ROOT_TOKEN=mzsQj8VU_zcSoSZsSX__
+  ita-by-ansible-legacy-role-vars-listup:
+    build:
+      dockerfile: ./ita_root/ita_by_ansible_legacy_role_vars_listup/Dockerfile
+      context: ../.
+      args:
+        - HTTP_PROXY=${HTTP_PROXY}
+        - http_proxy=${HTTP_PROXY}
+        - HTTPS_PROXY=${HTTPS_PROXY}
+        - https_proxy=${HTTPS_PROXY}
+    environment:
+      - EXECUTE_INTERVAL=1
+      - PYTHONPATH=/exastro/
+      - LANGUAGE=en
+      - USER_ID=20401
+      - DB_HOST=ita-mariadb
+      - DB_PORT=3306
+      - DB_DATADBASE=ITA_DB
+      - DB_USER=ITA_USER
+      - DB_PASSWORD=password
+      - DB_ROOT_PASSWORD=password
+      - STORAGEPATH=/storage/
+  ita-by-ansible-towermaster-sync:
+    build:
+      dockerfile: ./ita_root/ita_by_ansible_towermaster_sync/Dockerfile
+      context: ../.
+      args:
+        - HTTP_PROXY=${HTTP_PROXY}
+        - http_proxy=${HTTP_PROXY}
+        - HTTPS_PROXY=${HTTPS_PROXY}
+        - https_proxy=${HTTPS_PROXY}
+    environment:
+      - EXECUTE_INTERVAL=1
+      - PYTHONPATH=/exastro/
+      - LANGUAGE=en
+      - USER_ID=20102
+      - DB_HOST=ita-mariadb
+      - DB_PORT=3306
+      - DB_DATADBASE=ITA_DB
+      - DB_USER=ITA_USER
+      - DB_PASSWORD=password
+      - DB_ROOT_PASSWORD=password
+      - STORAGEPATH=/storage/
   ita-by-conductor-synchronize:
     build:
       dockerfile: ./ita_root/ita_by_conductor_synchronize/Dockerfile
@@ -149,9 +148,10 @@ services:
         - HTTPS_PROXY=${HTTPS_PROXY}
         - https_proxy=${HTTPS_PROXY}
     environment:
-      - EXECUTE_INTERVAL=10
+      - EXECUTE_INTERVAL=1
       - PYTHONPATH=/exastro/
-      - DEFAULT_LANGUAGE=en
+      - LANGUAGE=en
+      - USER_ID=30101
       - DB_HOST=ita-mariadb
       - DB_PORT=3306
       - DB_DATADBASE=ITA_DB
@@ -159,29 +159,24 @@ services:
       - DB_PASSWORD=password
       - DB_ROOT_PASSWORD=password
       - STORAGEPATH=/storage/
-      - GITLAB_HOST=10.197.18.183
-      - GITLAB_PORT=40080
-      - GITLAB_ROOT_TOKEN=mzsQj8VU_zcSoSZsSX__
-  # ita-by-menu-create:
-  #   build:
-  #     dockerfile: ./ita_root/ita_by_menu_create/Dockerfile
-  #     context: ../.
-  #     args:
-  #       - HTTP_PROXY=${HTTP_PROXY}
-  #       - http_proxy=${HTTP_PROXY}
-  #       - HTTPS_PROXY=${HTTPS_PROXY}
-  #       - https_proxy=${HTTPS_PROXY}
-  #   environment:
-  #     - EXECUTE_INTERVAL=10
-  #     - PYTHONPATH=/exastro/
-  #     - DEFAULT_LANGUAGE=en
-  #     - DB_HOST=ita-mariadb
-  #     - DB_PORT=3306
-  #     - DB_DATADBASE=ITA_DB
-  #     - DB_USER=ITA_USER
-  #     - DB_PASSWORD=password
-  #     - DB_ROOT_PASSWORD=password
-  #     - STORAGEPATH=/storage/
-  #     - GITLAB_HOST=10.197.18.183
-  #     - GITLAB_PORT=40080
-  #     - GITLAB_ROOT_TOKEN=mzsQj8VU_zcSoSZsSX__
+  ita-by-menu-create:
+    build:
+      dockerfile: ./ita_root/ita_by_menu_create/Dockerfile
+      context: ../.
+      args:
+        - HTTP_PROXY=${HTTP_PROXY}
+        - http_proxy=${HTTP_PROXY}
+        - HTTPS_PROXY=${HTTPS_PROXY}
+        - https_proxy=${HTTPS_PROXY}
+    environment:
+      - EXECUTE_INTERVAL=1
+      - PYTHONPATH=/exastro/
+      - LANGUAGE=en
+      - USER_ID=50101
+      - DB_HOST=ita-mariadb
+      - DB_PORT=3306
+      - DB_DATADBASE=ITA_DB
+      - DB_USER=ITA_USER
+      - DB_PASSWORD=password
+      - DB_ROOT_PASSWORD=password
+      - STORAGEPATH=/storage/
