+++ ita_root/ita_api_organization/libs/organization_common.py
@@ -32,8 +32,10 @@ def before_request_handler():
     """
     try:
         set_api_timestamp()
+
+        g.LANGUAGE = os.environ.get("DEFAULT_LANGUAGE")
         # create app log instance and message class instance
-        g.applogger = AppLog()
+        g.applogger = AppLog(g.LANGUAGE)
         g.appmsg = MessageTemplate()
 
         check_request_body()
@@ -65,11 +67,9 @@ def before_request_handler():
         # set language
         language = request.headers.get("Language")
         if language:
+            g.LANGUAGE = language
             g.appmsg.set_lang(language)
             g.applogger.info("LANGUAGE({}) is set".format(language))
-        else:
-            language = os.environ.get("DEFAULT_LANGUAGE")
-        g.LANGUAGE = language
 
         # initialize setting organization-db connect_info and connect check
         common_db = DBConnectCommon()  # noqa: F405
